backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.16.2
python-multipart==0.0.12
# PyTorch CPU versions (install first for caching, avoids CUDA bloat)
#torch==2.1.0+cpu torchvision==0.16.0+cpu torchaudio==2.1.0+cpu --index-url https://download.pytorch.org/whl/cpu
python-jose[cryptography]==3.3.0
qrcode[pil]==8.2
bcrypt==4.0.1
scikit-learn>=1.5.0
numpy>=2.0.0
spacy>=3.7.0
# axios
python-dotenv==1.0.0
redis==5.0.1
bleach==6.1.0
prometheus-client==0.19.0
psutil==5.9.6
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.25
mysql-connector-python==8.2.0
alembic==1.12.1
sentence-transformers==2.2.2
faiss-cpu
openai==1.3.0
xlsxwriter
pandas>=2.0.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-multipart>=0.0.6
web3>=6.11.0
eth-account>=0.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
